GS-10 서비스와 그와 연결된 서비스를 관리하는 객체 추가
####Change List
 - 로드밸런서 타입 제네릭 타입으로 변경
 - 연결되는 서비스를 Dictionary로 관리하도록 변경

+++ DB/Service/DBService.cs
@@ -52,7 +52,7 @@ namespace DB.Service
             {
                 Username = request.Username,
                 Message = "",
-                StatusCode = result,
+                StatusCode = (int)result,
             };
         }
